# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "The Nexus API",
    "version": "1.0.0",
    "description": "API documentation for The Nexus platform, standardized with v1 routes."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/professionals/upload": {
      "post": {
        "summary": "Upload a file for professional profile",
        "tags": [
          "Professionals",
          "Upload"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "type"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to upload (PDF, JPEG, PNG)"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "profile",
                      "resume"
                    ],
                    "description": "Type of file being uploaded"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "File uploaded successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "fileUrl": {
                          "type": "string",
                          "description": "Public URL of the uploaded file",
                          "example": "/uploads/professionals/user123_profile_abc123.jpg"
                        },
                        "fileName": {
                          "type": "string",
                          "description": "Original filename",
                          "example": "profile_photo.jpg"
                        },
                        "fileSize": {
                          "type": "number",
                          "description": "File size in bytes",
                          "example": 2048576
                        },
                        "fileType": {
                          "type": "string",
                          "description": "MIME type of the file",
                          "example": "image/jpeg"
                        },
                        "profileCompleteness": {
                          "type": "number",
                          "description": "Updated profile completeness percentage",
                          "example": 85
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file or missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "No file provided"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - professional access required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Forbidden: Professional access required"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "File too large",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "File size too large. Maximum 5MB allowed"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid file type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid file type. Only PDF, JPEG, JPG, and PNG files are allowed"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to upload file"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}